fix & address comment

+++ lib/Dialect/Torch/IR/TorchOps.cpp
@@ -185,7 +185,7 @@ static Value getScalarFloatValue(Value input, Location loc,
   if (auto valueTensorLiteralOp = input.getDefiningOp<ValueTensorLiteralOp>()) {
     auto val = valueTensorLiteralOp.getValue()
                    .cast<DenseElementsAttr>()
-                   .getSplatValue<float>();
+                   .getSplatValue<double>();
     return rewriter.create<Torch::ConstantFloatOp>(
         loc, rewriter.getF64FloatAttr(val));
   } else if (auto primNumToTensorScalarOp =
